requirements.txt
Django==2.2.1
djangorestframework==3.9.4
mysqlclient==1.4.2.post1